Add more detail around QuestionMediaManager implementation and move ViewModelAudioPlayer
