+++ collect_app/src/main/java/org/odk/collect/android/utilities/QuestionMediaManager.java
@@ -4,6 +4,11 @@ import androidx.annotation.Nullable;
 
 import java.io.File;
 
+/**
+ * Provides an interface for widgets to manage answer files. This lets them delete answer files
+ * when an answer is cleared, replace answer files when an answer is changed and also access answer
+ * files.
+ */
 public interface QuestionMediaManager {
 
     @Nullable
