+++ collect_app/src/main/java/org/odk/collect/android/widgets/utilities/ViewModelAudioPlayer.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.audio;
+package org.odk.collect.android.widgets.utilities;
 
 import androidx.lifecycle.LifecycleOwner;
 import androidx.lifecycle.Observer;
